feat: Reorganised repository to have a cleaner structure
