+++ frontend/src/chess/components/board/BoardConfig.ts
