+++ frontend/src/chess/components/board/Piece.tsx
