+++ frontend/src/chess/components/board/Square.tsx
