+++ frontend/src/chess/components/controls/GameMoveLogSidebar.tsx
