+++ frontend/src/chess/components/controls/GameOptionsSidebar.css
