+++ frontend/src/chess/components/controls/GameOptionsSidebar.tsx
