+++ frontend/src/chess/components/history/types.ts
