--- frontend/src/chess/coreMove.ts
@@ -1,2 +1,0 @@
-// Core game logic, including move generation, validation, and application
-// This file would handle the core move logic for the chess game, including special moves as well.
