--- frontend/src/chess/position.ts
@@ -1,1 +1,0 @@
-// board representation, FEN load/save
